Changing to a different CellML model with SED-ML to control simulation parameters.
The Noble model is a bit confused as it deals with time in seconds, making it non-intuitive to use. Switched to ORd model for testing and it runs and gives a reasonable output. Using a SED-ML document as the input so that the simulation settings can be controlled directly. This docker container will build, but the expected 'run' method is failing for me. Running a terminal in the container the simulation executes as expected, but the 'docker run...' fails for some reason.